fix: electron api type frontend access
The electron api type integrated widow is not reflecting in the frontend.

so made separate electron.-api.ts file which contains types which were in preload.ts

created electron.d.ts to integrate electronApi type to global window
