+++ src/types/electron.d.ts
@@ -1,0 +1,7 @@
+declare global {
+  interface Window {
+    electronAPI: ElectronAPI
+  }
+}
+
+export {}
